# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RoadSoS Emergency Response API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SosAlertType = typeof SosAlertType[keyof typeof SosAlertType];


export const SosAlertType = {
  accident: 'accident',
  breakdown: 'breakdown',
  medical: 'medical',
  fire: 'fire',
  other: 'other',
} as const;

export type SosAlertStatus = typeof SosAlertStatus[keyof typeof SosAlertStatus];


export const SosAlertStatus = {
  active: 'active',
  resolved: 'resolved',
  cancelled: 'cancelled',
} as const;

export interface SosAlert {
  id: string;
  userId: string;
  type: SosAlertType;
  status: SosAlertStatus;
  latitude: number;
  longitude: number;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export type SosAlertInputType = typeof SosAlertInputType[keyof typeof SosAlertInputType];


export const SosAlertInputType = {
  accident: 'accident',
  breakdown: 'breakdown',
  medical: 'medical',
  fire: 'fire',
  other: 'other',
} as const;

export interface SosAlertInput {
  userId: string;
  type: SosAlertInputType;
  latitude: number;
  longitude: number;
  address?: string;
  description?: string;
}

export type SosAlertUpdateStatus = typeof SosAlertUpdateStatus[keyof typeof SosAlertUpdateStatus];


export const SosAlertUpdateStatus = {
  active: 'active',
  resolved: 'resolved',
  cancelled: 'cancelled',
} as const;

export interface SosAlertUpdate {
  status?: SosAlertUpdateStatus;
  description?: string;
}

export interface EmergencyContact {
  id: string;
  userId: string;
  name: string;
  phone: string;
  relation: string;
  /** @nullable */
  photoUrl?: string | null;
}

export interface EmergencyContactInput {
  userId: string;
  name: string;
  phone: string;
  relation: string;
  photoUrl?: string;
}

export type IncidentType = typeof IncidentType[keyof typeof IncidentType];


export const IncidentType = {
  accident: 'accident',
  roadblock: 'roadblock',
  flood: 'flood',
  fire: 'fire',
  hazard: 'hazard',
} as const;

export type IncidentSeverity = typeof IncidentSeverity[keyof typeof IncidentSeverity];


export const IncidentSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface Incident {
  id: string;
  type: IncidentType;
  latitude: number;
  longitude: number;
  severity: IncidentSeverity;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  reportedBy?: string | null;
  createdAt: string;
}

export type IncidentInputType = typeof IncidentInputType[keyof typeof IncidentInputType];


export const IncidentInputType = {
  accident: 'accident',
  roadblock: 'roadblock',
  flood: 'flood',
  fire: 'fire',
  hazard: 'hazard',
} as const;

export type IncidentInputSeverity = typeof IncidentInputSeverity[keyof typeof IncidentInputSeverity];


export const IncidentInputSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface IncidentInput {
  type: IncidentInputType;
  latitude: number;
  longitude: number;
  severity: IncidentInputSeverity;
  description?: string;
  reportedBy?: string;
}

export interface DashboardSummary {
  totalAlerts: number;
  activeAlerts: number;
  resolvedAlerts: number;
  totalContacts: number;
  recentIncidents: number;
}

export type ListSosAlertsParams = {
userId?: string;
limit?: number;
};

export type ListContactsParams = {
userId: string;
};

export type ListIncidentsParams = {
lat?: number;
lng?: number;
/**
 * radius in km
 */
radius?: number;
};

export type GetDashboardSummaryParams = {
userId: string;
};