# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RoadSoS Emergency Response API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalAlerts: number;
  activeAlerts: number;
  resolvedAlerts: number;
  totalContacts: number;
  recentIncidents: number;
}