# lib/api-zod/src/generated/types/emergencyContactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RoadSoS Emergency Response API
 * OpenAPI spec version: 0.1.0
 */

export interface EmergencyContactInput {
  userId: string;
  name: string;
  phone: string;
  relation: string;
  photoUrl?: string;
}