# lib/api-zod/src/generated/types/incidentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RoadSoS Emergency Response API
 * OpenAPI spec version: 0.1.0
 */
import type { IncidentInputSeverity } from './incidentInputSeverity';
import type { IncidentInputType } from './incidentInputType';

export interface IncidentInput {
  type: IncidentInputType;
  latitude: number;
  longitude: number;
  severity: IncidentInputSeverity;
  description?: string;
  reportedBy?: string;
}