# lib/api-zod/src/generated/types/sosAlertInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RoadSoS Emergency Response API
 * OpenAPI spec version: 0.1.0
 */
import type { SosAlertInputType } from './sosAlertInputType';

export interface SosAlertInput {
  userId: string;
  type: SosAlertInputType;
  latitude: number;
  longitude: number;
  address?: string;
  description?: string;
}