# lib/api-zod/src/generated/types/sosAlertUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RoadSoS Emergency Response API
 * OpenAPI spec version: 0.1.0
 */
import type { SosAlertUpdateStatus } from './sosAlertUpdateStatus';

export interface SosAlertUpdate {
  status?: SosAlertUpdateStatus;
  description?: string;
}